+++ NEWS.md
@@ -1,11 +1,11 @@
 V 0.3.4
 ========
-- Improvement of function whichAreBijection. It is 2 to 3 times faster than previous version.
+- Improvement of function whichAreBijection. It is 2 to 15 time faster than previous version.
 - Bug fixes:
 	- *generateFactorFromDate*: default value was missing. Fixed.
 - New features:
 	- New features in existing functions:
-		- *fastFilterVariables* has a new parameter (level) to choose which types of variables you want to drop
+		- *fastFilterVariables* has a new parameter (level) to choose which types of filtering to perform
 V 0.3.3
 ========
 - New features: